remove need for XS module
This is a fork of Philippe M. Chiasson's Apache2-Layout-0.6, which was an
apache2 version of the old apache1 mod_layout.

The XS code was segfaulting at run-time, but it's not really necessary at all
since this bug was fixed in mod_perl 2.0.4 (r607687)
https://mail-archives.apache.org/mod_mbox/perl-modperl-cvs/200712.mbox/%3C20071231064858.EA2BF1A9842@eris.apache.org%3E

From mod_perl's Changelog
2.0.4 April 16, 2008
   Fix a crash when running a sub-request from within a filter where mod_perl
   was not the content handler. [Gozer]

If this tests out ok I'll see about resubmitting it upstream to cpan.